# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FastAppoint API",
    "description": "Backend API for FastAppoint - Booking management system",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/login-success": {
      "get": {
        "tags": [
          "Authentication"
        ],
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}